requirements.txt
# Web framework
Django
Pillow
django-filter
django-quill-editor
djangorestframework

# Task queue
celery
redis
celery[redis]

# Database
mysqlclient

# Localization
django-modeltranslation
pytz

# Authentication and account management
django-allauth
django-allauth[socialaccount]

# Environment variables
python-dotenv

# Code verification tools
flake8
flake8-import-order
flake8-docstrings
mypy
